+++ src/response.lisp
@@ -14,6 +14,8 @@
   (:import-from #:weblocks/commands
                 #:add-command)
   (:import-from #:quri)
+  (:import-from #:alexandria
+                #:assoc-value)
   (:export #:immediate-response
            #:make-response
            #:add-header
@@ -26,7 +28,8 @@
            #:get-code
            #:get-headers
            #:get-custom-headers
-           #:get-content-type))
+           #:get-content-type
+           #:add-retpath-to))
 (in-package weblocks/response)
 
 
@@ -121,6 +124,32 @@
          (parsed-new-path (quri:uri new-path))
          (new-url (quri:merge-uris parsed-new-path
                                    parsed-base)))
+    (quri:render-uri new-url)))
+
+
+(defun add-retpath-to (uri &key (retpath (weblocks/request:get-uri)))
+  "Adds a \"retpath\" GET parameter to the giving URL.
+
+   Keeps all other parameters and overwrites \"retpath\" parameter if it is
+   already exists in the URL.
+
+   By default, retpath is the current page, rendered by the weblocks.
+   This is very useful to redirect user to login page and return him to the
+   same page where he has been before."
+  (let* ((parsed-base (quri:uri uri))
+         (query (quri:uri-query parsed-base))
+         (parsed-query (when query
+                         (quri:url-decode-params query)))
+         (_ (setf (assoc-value parsed-query
+                                          "retpath"
+                                          :test 'string-equal)
+                  retpath))
+         (new-query (quri:url-encode-params parsed-query))
+         (parsed-new-path (quri:uri (concatenate 'string "?"
+                                                 new-query)))
+         (new-url (quri:merge-uris parsed-new-path
+                                   parsed-base)))
+    (declare (ignorable _))
     (quri:render-uri new-url)))